allow 'optional' modules in the Vermont module graph
if a module calls Source::setAllowOptionalModules() in the constructor,
Vermont's module connector will allow succeeding modules with
non-compatible element types, so that one module may send packets AND
IPFIX flows to succeeding modules. If the module calls send() to send
elements to succeeding modules, only the default element type is supported.
To send other element types to succeeding modules, the module needs to use
functions Source::getSucceedingOptModuleCount() and
Source::getSuccedingOptModuleInstance() to detect optional modules (they
need to be checked for validity!), and send elements to them.