HTTP aggregation enhancement for the PacketAggregator module
The funcionality of the PacketAggregator will be extended to support the aggregation of HTTP based communication.
With this extension it will be possible to create IPFIX flows based on HTTP request and response messages which belong to each other.

This initial commit provides a rudimentary version of the enhancement.

Additional changes:

- minor bugfixes for DPA aggregation in the PacketHashtable
- fix version.c for proper kernel version string handling. since kernel 3.x version strings are shorter
changes were taken from:
http://anonscm.debian.org/gitweb/?p=collab-maint/procps.git;a=tree;f=proc;h=688f1c3788205bcab1707ae3c481b7da06fe0443;hb=HEAD