Skip non-TCP packets + Fixed faulty aggregation
Non-TCP packets are skipped as HTTP aggregation and TCP connection
handling require and only work properly with TCP packets.

The way how packets/tcp streams were match to their according bucket
was faulty. In some cases packets were aggregated into the wrong
bucket. This issue was fixed by adding a condition when checking
for flow quality: Not only the network 4-tuple has to match to
determine the right bucket, but also the stream id, which identifies
the TCP connection, has to match.