prevent buffer overflow + bug fixes
included two new config options to control the size of the used buffers:

- tcpmonBufferSize:
The maximal number of bytes buffered per TCP connection. If segments are out-of-order we have to buffer TCP packets until the gaps are filled up. The total packet lengths are considered.
defaults to 1048576 (1 MiB)

- httpaggBufferSize:
The maximal number of bytes buffered per HTTP message. If the parser cannot fully parse the payload of TCP segment, we have to buffer the payload and combine it with the payload of the next segment. This value defines the maximum number of bytes of payload which can be combined/buffered.
defaults to 10240 (10 KiB)

Fix:
- Multiple HTTP responses per segment weren't properly aggregated into existing flows
- HTTP status was set faulty in the case of a HTTP protocol switch