Fixed another bucket collision issue
The hash function used by the packet aggregator uses a 32bit hash function to generate a hash. This resulting hash value gets bitwise AND combined with the bit-'size' of the hashtable. Default setting is to use a bitsize value of 17 for the hashtable ( 0x40000 buckets) and hence only 17bit of the resulting hash value are used. That obviously causes collisions, as almost half of the hash value gets cut off. Allthough we store the ID of the TCP connection in each bucket and compare against it, to avoid collisions, we still suffer from collisions if within a TCP connection two HTTP messages get the same hash value, what seems to be the case at very high packet rates. So we additionally check if the flowID (HTTP flow count number), which now also is stored in the bucket, is appropriate.