Allow HTTP aggregation without storing payload
Previously the length of the fields frontPayload and revFrontPayload had
to be bigger than 5. HTTP aggregation requires these fields to be present
in the configuration, but payload aggregation is not always desired.

Thus, from now on it is possible to configure these fields with a length
of 0, which prevents aggregating the payload into the IPFIX record.